bytes: conslidate bytes sources into iobuf
Having a separate compilation unit for a dozen or so lines of code is
wasteful in compilation cost and on-disk space, especially for our
static build where on-disk object files are often >1MB when including
seastar headers.
