+++ src/v/bytes/CMakeLists.txt
@@ -2,7 +2,6 @@ v_cc_library(
   NAME
     bytes
   SRCS
-    "bytes.cc"
     "iobuf.cc"
     oncore.cc
   DEPS
